+++ ui/utils/messageline.h
@@ -59,6 +59,8 @@ public:
     void setPalAvatar(const QString& jid, const QString& path);
     void dropPalAvatar(const QString& jid);
     void changeMessage(const QString& id, const QMap<QString, QVariant>& data);
+    void setExPalAvatars(const std::map<QString, QString>& data);
+    void movePalAvatarToEx(const QString& name);
     
 signals:
     void resize(int amount);
@@ -90,6 +92,7 @@ private:
     std::map<QString, Index> palMessages;
     std::map<QString, QString> uploadPaths;
     std::map<QString, QString> palAvatars;
+    std::map<QString, QString> exPalAvatars;
     QVBoxLayout* layout;
     
     QString myName;